mysite/requirements.txt
asgiref==3.2.3
Django==3.0.4
django-nested-admin==3.2.4
python-monkey-business==1.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
